requirements.txt
black
flask
flask-executor
flask-sqlalchemy
pre-commit
slackclient
tabulate
